+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 > _Please don't set a new version if you are the first to make the section for `master`._
 
 * Improves the inline docs on `danger local` and `danger pr`. [@orta](https://github.com/orta)
+* Improves the GitLab CI error handling if port number is accidentally included in host. [@mbogh](https://github.com/mbogh)
 
 ## 5.5.5